requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2

